fix: I-02 consistent error handling - add ComputeHashesWithReporting
- Added ChdHashComputeResult record type for error-code-based results
- Added ComputeHashesWithReporting method that returns ChdError codes
- ComputeHashes now delegates to ComputeHashesWithReporting and throws on error
- Added HashRegionWithReporting helper that returns error codes
- Minor formatting fixes (blank lines after guards)

+++ CHDSharpEncoder/ChdEncoder.cs
@@ -1307,7 +1307,7 @@ public static class ChdEncoder
     private static int ReadCdHunk(uint hunkIndex, byte[] buffer, CdToc toc, int framesPerHunk, ulong totalFrames,
         Dictionary<string, FileStream> files)
     {
-        var hunkStartFrame = (long)hunkIndex * framesPerHunk;
+        var hunkStartFrame = hunkIndex * framesPerHunk;
         for (var f = 0; f < framesPerHunk; f++)
         {
             var frame = hunkStartFrame + f;
